configure.ac: check for asciidoc and xmlto
asciidoc and xmlto are now needed to build the lttng(1) man pages from
AsciiDoc sources. This is only checked if the --enable-build-man-pages
is passed to ./configure; otherwise, the man pages are still installed
if their troff version exists, which should always be the case in
tarball distributions so that the user does not need asciidoc and xmlto.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>