sessiond: use `loglevel_value` and `loglevel_type` names
By using the `loglevel_value` and `loglevel_type` names instead
of `loglevel` for one or the other, some unsettling
inconsistencies are exposed.

This patch only changes the names to show the weird stuff, e.g.:

    key.loglevel_type = loglevel_value;

A future patch will fix this.

The only `loglevel` names left untouched are those in public headers
as well as those in the tools<->UST ABI.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>