Fix: lttng-crash: don't segfault on -e flag
The optstring parameter given to `getopt_long` was such that when
passing the -e flag to specify a viewer other than the default,
`optarg` would remain NULL and cause a segfault when passing it to
`strdup`. A colon after the flag in the optstring makes the function
expect a following argument, in this case the path to the viewer.

Signed-off-by: Antoine Busque <abusque@efficios.com>