Fix: Log and ignore SIGINT and SIGTERM in run_as worker
The run_as worker is in the same process group as its parent and
will receive both SIGINT and SIGTERM. However, we want to give
the worker a chance to tear itself down gracefully when its
parent closes the command socket.

The run_as worker will now ignore these signals (although it will
log them) and wait for the parent to induce the teardown.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>