Select which binaries/extras to build at configure time
Introduce new flags to the configure script to select what to build:

* By default everything is built *

--disable-bin-lttng
    Do not build anything under src/bin/lttng.
--disable-bin-lttng-consumerd
    Do not build anything under src/bin/lttng-consumerd.
--disable-bin-lttng-crash
    Do not build anything under src/bin/lttng-crash.
--disable-bin-lttng-relayd
    Do not build anything under src/bin/lttng-relayd.
--disable-bin-lttng-sessiond
    Do not build anything under src/bin/lttng-sessiond.
--disable-extras
    Do not build anything under extras.

Only the necessary libs (from src/lib and src/common) are built.
libcommon and libconfig are always built since they are used by every
target.

If any lttng-* binaries are disabled, 'tests' will not be built.
On ./configure a warning is shown to the user.
Also 'make check' and 'make checkinstall' only return a warning when 'tests'
are not built.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>