Fix: load event state (enabled/disabled) correctly
This bug fix is a workaround due to limitations of lttng_disable_event_ext
regarding the disabling of events with similar name but different
characteristics. Although lttng_disable_event_ext provides support for
disabling by name and filter string it does not support exclusion.

The loading of events is cut in 3 phases.

1 - Create all events regardless of their state.
2 - Disable all events.
3 - Enable only the events with the 'enabled' state.

Fixes #959

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>