Fix: Set loopback adress in set_ip_addr if gethostbyname2 fails
Some systems may not have "localhost" defined in accordance with IETF
RFC 6761. According to this RFC, applications may recognize
"localhost" names as special and resolve to the appropriate loopback
address.

We choose to use the system name resolution API first to honor its
network configuration. If this fails, we resolve to the appropriate
loopback address. This is done to accomodate systems which may want to
start tracing before their network configured.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>