configure.ac: beautify report
This beautifies the ./configure end report by aligning all the
fields and adding colors when using an interactive shell
with color support and with a standard output stream connected
to a terminal.

It also uses /bin/fold, if available, to wrap the lines of the
beer description instead of printing a very long line.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>