Fix: convey enum value signedness into metadata
Currently, passing an enum range of:

    ctf_enum_range("blah", 0, UINT_MAX)

in LTTng-UST will print a range of 0 ... -1 in the generated CTF
metadata, which does not reflect signedness of the values.

Also, struct ustctl_enum_entry is missing a LTTNG_PACKED attribute,
which is against our protocol rules.

This change needs to be pushed in locked-step into lttng-tools and
lttng-ust, since it breaks the protocol between the two when UST
uses the new enumeration type (introduced in 2.8.0-rc1).

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>