Fix: do not return error on LTTNG_ERR_SNAPSHOT_NODATA
A warning is fine since the user has no control on
whether or not applications (or the kernel) have
produced any event between the start of the tracing
session and the recording of the snapshot.

MI wise the command is not a success since nothing was
recorded. The command line return code is CMD_SUCCESS.

refs #1002

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>