Fix: Add missing includes needed in lttng-kernel.h
LTTNG_SYMBOL_NAME_LEN is defined in lttng/constant.h
enum lttng_event_output is defined in lttng/event.h

This is not causing any trouble at the moment because constant.h and
event.h are also included through other includes when lttng-kernel.h is
included. Explicitely including the header files when used makes it
cleaner and easier to debug.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>